Changes required for jenkins
It turns out there were some changes needed to make the script
work when run from jenkins.

1. For some reason, the find command used was the one from Windows, not the unix one.
Even if on the same machine the correct one was used when the script was run manually
from CLI. So to make this more reliable, the path for cygwin is now hardcoded on Windows.

2. vagrant destroy is now forced - before this change, vagrant destroy required
confirmation on the console. This would not work on Jenkins. And it would defeat
the purpose of automated builds.

3. A symlink to the latest installation directory is created at the end of the setup
 - latest-installation. This way the subsequent integration test can always use this
one fixed path for cdk.